# openapi.yaml
openapi: 3.0.3
info:
  title: Sample Node API
  description: A simple product catalog REST API — demo service for Meridian EA governance
  version: 1.0.0
  contact:
    name: Platform Engineering
servers:
  - url: http://localhost:3001
    description: Local development
paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  uptime:
                    type: number
                    example: 42.3

  /api/products:
    get:
      summary: List all products
      operationId: listProducts
      tags: [Products]
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category name
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
    post:
      summary: Create a product
      operationId: createProduct
      tags: [Products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/products/{id}:
    get:
      summary: Get a product by ID
      operationId: getProduct
      tags: [Products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Product:
      type: object
      required: [id, name, price, category]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Laptop Pro
        price:
          type: number
          format: float
          example: 1299.99
        category:
          type: string
          example: Electronics

    ProductInput:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string

    ProductList:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer

    Error:
      type: object
      properties:
        error:
          type: string